fix incorrect redirection for PA start page
adjust app store payload helper to have both pa and nsm versions
use correct params for submission confirmation controller
