+++ spec/system/nsm/claim_confirmation_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe 'User can see an application status', :stub_app_store_search, :st
   let(:claim) { create(:claim, :complete, :case_type_breach, ufn: '12122025/001', state: :submitted) }
 
   before do
-    stub_app_store_payload(claim, :submitted, 'LAA-ABC321')
+    stub_nsm_app_store_payload(claim, :submitted, 'LAA-ABC321')
     visit provider_entra_id_omniauth_callback_path
     visit nsm_steps_claim_confirmation_path(claim.id)
   end
